Refactor: rename components to correct Formate
